add tests for ScoreBoardActivity refactor PlayerAdapterTest.kt and ScoreBoardActivityTest.kt
