+++ app/src/androidTest/java/ch/sdp/vibester/scoreboard/PlayerAdapterTest.kt
@@ -1,4 +1,4 @@
-package ch.sdp.vibester
+package ch.sdp.vibester.scoreboard
 
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
