+++ app/src/androidTest/java/ch/sdp/vibester/scoreboard/ScoreBoardActivityTest.kt
@@ -1,22 +1,18 @@
-package ch.sdp.vibester
+package ch.sdp.vibester.scoreboard
 
 import androidx.recyclerview.widget.RecyclerView
-import androidx.test.core.app.ActivityScenario
+import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.rules.ActivityScenarioRule
-import org.junit.After
-import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.ViewInteraction
 import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.contrib.RecyclerViewActions
-import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
-import androidx.test.espresso.matcher.ViewMatchers.withId
+import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import ch.sdp.vibester.scoreboard.ScoreBoardActivity
+import ch.sdp.vibester.R
 
 @RunWith(AndroidJUnit4::class)
 class ScoreBoardActivityTest {
@@ -24,10 +20,6 @@ class ScoreBoardActivityTest {
     @Rule
     @JvmField
     val activityRule = ActivityScenarioRule(ScoreBoardActivity::class.java)
-
-    @Before
-    fun setUp() {
-    }
 
     @Test
     fun recycleViewToViewTest() {
@@ -40,8 +32,19 @@ class ScoreBoardActivityTest {
             .perform(RecyclerViewActions.actionOnItemAtPosition<RecyclerView.ViewHolder>(2, click()))
     }
 
-    @After
-    fun tearDown() {
+    @Test
+    fun recycleViewScrollDownTest() {
+        val recyclerView = RecyclerView(ApplicationProvider.getApplicationContext())
+        val itemCount = recyclerView.adapter?.itemCount
+        if(itemCount != null) {
+            onView(withId(R.id.recycler_view)).perform(RecyclerViewActions.scrollToPosition<RecyclerView.ViewHolder>(itemCount.minus(1)))
+        }
+    }
 
+    @Test
+    fun recycleViewShowItemTest() {
+        onView(withId(R.id.recycler_view)).perform(RecyclerViewActions.actionOnItemAtPosition<RecyclerView.ViewHolder>(5, click()))
+        val nameItem = "Cinnamon Roll"
+        onView(withText(nameItem)).check(matches(isDisplayed()))
     }
 }
